// src/modules/config/config.types.ts
export enum ENV_TYPES {
  development = 'development',
  production = 'production',
}

export enum DB_PROPS {
  TYPEORM_CONNECTION = 'TYPEORM_CONNECTION',
  TYPEORM_HOST = 'TYPEORM_HOST',
  TYPEORM_USERNAME = 'TYPEORM_USERNAME',
  TYPEORM_PASSWORD = 'TYPEORM_PASSWORD',
  TYPEORM_DATABASE = 'TYPEORM_DATABASE',
  TYPEORM_PORT = 'TYPEORM_PORT',
  TYPEORM_SYNCHRONIZE = 'TYPEORM_SYNCHRONIZE',
  TYPEORM_LOGGING = 'TYPEORM_LOGGING',
}

export enum OMDB_PROPS {
  OMDB_API_URL = 'OMDB_API_URL',
  OMDB_API_KEY = 'OMDB_API_KEY',
}